refactor(gateway): extract platform layer (state + static_files) — ironclaw#2599 stage 1
First increment of the ironclaw#2599 gateway platform/feature split. Moves
the shared gateway state and the unauthenticated static-serving surface out
of the 8.5k-line `server.rs` monolith into a new `platform/` subtree:

- `platform/state.rs` — `GatewayState`, `RateLimiter`, `PerUserRateLimiter`,
  `WorkspacePool` (+ `WorkspaceResolver` impl), `FrontendHtmlCache`,
  `FrontendCacheKey`, `ActiveConfigSnapshot`, `PromptQueue`,
  `RoutineEngineSlot`, `rate_limit_key_from_headers`.
- `platform/static_files.rs` — CSP directive set + `BASE_CSP_HEADER`
  (single source of truth for both the global header layer and the
  per-response nonce variant), `build_frontend_html` + cache-key plumbing,
  unauthenticated static handlers (`/`, `/style.css`, `/app.js`,
  `/theme.css`, `/favicon.ico`, `/i18n/*`, `/admin*`, `/api/health`), and
  the authenticated `/projects/{id}/...` file-serving routes with the
  ownership check.

`server.rs` keeps `start_server()`, the route table, and the feature
handlers that have not yet moved (OAuth callbacks, chat, extensions,
pairing, logs, gateway status). It re-exports the relocated state types
under their old paths so the 30+ external call sites that reach for
`crate::channels::web::server::GatewayState` continue to resolve without
churn — follow-up PRs update them incrementally.

`CLAUDE.md` file map is updated to document `platform/` vs the
transitional `handlers/` folder and explains the layering rule:
features depend on platform, never the reverse.

No behavior change — route table, CSP policy, cache semantics, and
multi-tenant guard rails are byte-identical. `cargo clippy --all
--benches --tests --examples --all-features` is clean; the relocated
tests (`test_base_csp_header_matches_build_csp_none`,
`test_stamp_nonce_into_html_*`, `workspace_pool_resolve_seeds_new_user_workspace`,
`test_build_frontend_html_returns_none_in_multi_tenant_mode`, etc.) pass
against the re-exported surface.

Scope is intentionally narrow to avoid colliding with the open
XL PRs that touch `server.rs` (#2548 workspace entities, #2555 tool
output, #2532 engine v2 sidebar) — this PR does not modify anything
inside the feature-handler blocks those PRs edit.

Stats: server.rs 8534 → 7463 lines (−1071); new files 1210 lines.
