+++ src/channels/web/mod.rs
@@ -20,6 +20,7 @@ pub mod log_layer;
 pub mod oauth;
 pub(crate) mod onboarding;
 pub mod openai_compat;
+pub(crate) mod platform;
 pub mod responses_api;
 pub mod server;
 pub mod sse;
