+++ src/channels/web/server.rs
@@ -1,11 +1,15 @@
 //! Axum HTTP server for the web gateway.
 //!
-//! Handles all API routes: chat, memory, jobs, health, and static file serving.
+//! Owns `start_server()` and the feature handlers that have not yet moved
+//! to domain modules. The platform-level pieces (shared state, rate
+//! limiters, the CSP/static/projects handlers) now live under
+//! `crate::channels::web::platform::*`; this file re-exports them so the
+//! existing `crate::channels::web::server::*` paths continue to resolve
+//! while the ironclaw#2599 migration is in progress.
 
 use std::convert::Infallible;
 use std::net::SocketAddr;
 use std::sync::Arc;
-use std::sync::atomic::{AtomicU64, Ordering};
 
 use axum::{
     Json, Router,
@@ -13,14 +17,14 @@ use axum::{
     http::{StatusCode, header},
     middleware,
     response::{
-        IntoResponse, Response,
+        IntoResponse,
         sse::{Event, KeepAlive, Sse},
     },
     routing::{get, post, put},
 };
 use serde::Deserialize;
 use sha2::{Digest, Sha256};
-use tokio::sync::{mpsc, oneshot};
+use tokio::sync::oneshot;
 use tokio_stream::StreamExt;
 use tower_http::cors::{AllowHeaders, CorsLayer};
 use tower_http::set_header::SetResponseHeaderLayer;
@@ -28,12 +32,9 @@ use uuid::Uuid;
 
 use axum::http::HeaderMap;
 
-use crate::agent::SessionManager;
-use crate::bootstrap::ironclaw_base_dir;
-use crate::channels::IncomingMessage;
 use crate::channels::relay::DEFAULT_RELAY_NAME;
 use crate::channels::web::auth::{
-    AdminUser, AuthenticatedUser, CombinedAuthState, UserIdentity, auth_middleware,
+    AdminUser, AuthenticatedUser, CombinedAuthState, auth_middleware,
 };
 use crate::channels::web::handlers::chat::chat_events_handler;
 use crate::channels::web::handlers::engine::{
@@ -71,35 +72,33 @@ use crate::channels::web::handlers::settings::{
 use crate::channels::web::handlers::skills::{
     skills_install_handler, skills_list_handler, skills_remove_handler, skills_search_handler,
 };
-use crate::channels::web::log_layer::LogBroadcaster;
-use crate::channels::web::sse::SseManager;
+use crate::channels::web::platform::static_files::{
+    BASE_CSP_HEADER, admin_css_handler, admin_html_handler, admin_js_handler, css_handler,
+    favicon_handler, health_handler, i18n_app_handler, i18n_en_handler, i18n_index_handler,
+    i18n_ko_handler, i18n_zh_handler, index_handler, js_handler, project_file_handler,
+    project_index_handler, project_redirect_handler, theme_css_handler, theme_init_handler,
+};
 use crate::channels::web::types::*;
 use crate::channels::web::util::{
     build_turns_from_db_messages, collect_generated_images_from_tool_results,
     enforce_generated_image_history_budget, tool_error_for_display, tool_result_preview,
     web_incoming_message,
 };
-use crate::db::Database;
-use crate::extensions::ExtensionManager;
 use crate::extensions::naming::extension_name_candidates;
-use crate::orchestrator::job_manager::ContainerJobManager;
 use crate::secrets::SecretConsumeResult;
-use crate::tools::ToolRegistry;
-use crate::workspace::Workspace;
 
-/// Shared prompt queue: maps job IDs to pending follow-up prompts for Claude Code bridges.
-pub type PromptQueue = Arc<
-    tokio::sync::Mutex<
-        std::collections::HashMap<
-            uuid::Uuid,
-            std::collections::VecDeque<crate::orchestrator::api::PendingPrompt>,
-        >,
-    >,
->;
-
-/// Slot for the routine engine, filled at runtime after the agent starts.
-pub type RoutineEngineSlot =
-    Arc<tokio::sync::RwLock<Option<Arc<crate::agent::routine_engine::RoutineEngine>>>>;
+// --- Backward-compat re-exports for the ironclaw#2599 migration ---
+//
+// The platform-level state types and rate limiters moved to
+// `crate::channels::web::platform::state`. External callers (handlers,
+// integration tests, `src/main.rs`, `src/app.rs`) still reach them via
+// `crate::channels::web::server::*`; re-export until the follow-up PR
+// updates every call site.
+pub(crate) use crate::channels::web::platform::state::rate_limit_key_from_headers;
+pub use crate::channels::web::platform::state::{
+    ActiveConfigSnapshot, FrontendCacheKey, FrontendHtmlCache, GatewayState, PerUserRateLimiter,
+    PromptQueue, RateLimiter, RoutineEngineSlot, WorkspacePool,
+};
 
 fn redact_oauth_state_for_logs(state: &str) -> String {
     let digest = Sha256::digest(state.as_bytes());
@@ -109,426 +108,6 @@ fn redact_oauth_state_for_logs(state: &str) -> String {
         let _ = write!(&mut short_hash, "{byte:02x}");
     }
     format!("sha256:{short_hash}:len={}", state.len())
-}
-
-pub(crate) fn rate_limit_key_from_headers(headers: &HeaderMap) -> String {
-    // Try X-Forwarded-For first (reverse proxy), then X-Real-IP.
-    let xff = headers
-        .get("x-forwarded-for")
-        .and_then(|v| v.to_str().ok())
-        .into_iter()
-        .flat_map(|s| s.split(','))
-        .map(str::trim)
-        .find_map(|candidate| candidate.parse::<std::net::IpAddr>().ok())
-        .map(|ip| ip.to_string());
-
-    if let Some(ip) = xff {
-        return ip;
-    }
-
-    headers
-        .get("x-real-ip")
-        .and_then(|v| v.to_str().ok())
-        .and_then(|s| s.trim().parse::<std::net::IpAddr>().ok())
-        .map(|ip| ip.to_string())
-        .unwrap_or_else(|| "unknown".to_string())
-}
-
-/// Simple sliding-window rate limiter.
-///
-/// Tracks the number of requests in the current window. Resets when the window expires.
-pub struct RateLimiter {
-    /// Requests remaining in the current window.
-    remaining: AtomicU64,
-    /// Epoch second when the current window started.
-    window_start: AtomicU64,
-    /// Maximum requests per window.
-    max_requests: u64,
-    /// Window duration in seconds.
-    window_secs: u64,
-}
-
-impl RateLimiter {
-    pub fn new(max_requests: u64, window_secs: u64) -> Self {
-        Self {
-            remaining: AtomicU64::new(max_requests),
-            window_start: AtomicU64::new(
-                std::time::SystemTime::now()
-                    .duration_since(std::time::UNIX_EPOCH)
-                    .unwrap_or_default()
-                    .as_secs(),
-            ),
-            max_requests,
-            window_secs,
-        }
-    }
-
-    /// Try to consume one request. Returns `true` if allowed, `false` if rate limited.
-    ///
-    /// Note: There is a benign TOCTOU race between checking `window_start` and
-    /// resetting it — two concurrent threads may both see an expired window
-    /// and reset it, granting a few extra requests at the window boundary.
-    /// This is acceptable for chat rate limiting where approximate enforcement
-    /// is sufficient, and avoids the cost of a Mutex.
-    pub fn check(&self) -> bool {
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs();
-
-        let window = self.window_start.load(Ordering::Relaxed);
-        if now.saturating_sub(window) >= self.window_secs {
-            // Window expired, reset
-            self.window_start.store(now, Ordering::Relaxed);
-            self.remaining
-                .store(self.max_requests - 1, Ordering::Relaxed);
-            return true;
-        }
-
-        // Try to decrement remaining
-        loop {
-            let current = self.remaining.load(Ordering::Relaxed);
-            if current == 0 {
-                return false;
-            }
-            if self
-                .remaining
-                .compare_exchange_weak(current, current - 1, Ordering::Relaxed, Ordering::Relaxed)
-                .is_ok()
-            {
-                return true;
-            }
-        }
-    }
-}
-
-/// Snapshot of the active (resolved) configuration exposed to the frontend.
-#[derive(Debug, Clone, Default, serde::Serialize)]
-pub struct ActiveConfigSnapshot {
-    pub llm_backend: String,
-    pub llm_model: String,
-    pub enabled_channels: Vec<String>,
-}
-
-/// Per-user rate limiter that maintains a separate sliding window per user_id.
-///
-/// Prevents one user from exhausting the rate limit for all users in multi-tenant mode.
-pub struct PerUserRateLimiter {
-    limiters: std::sync::Mutex<lru::LruCache<String, RateLimiter>>,
-    max_requests: u64,
-    window_secs: u64,
-}
-
-impl PerUserRateLimiter {
-    // SAFETY: 2048 is non-zero, so the unwrap in `new()` is infallible.
-    const MAX_KEYS: std::num::NonZeroUsize = match std::num::NonZeroUsize::new(2048) {
-        Some(v) => v,
-        None => unreachable!(),
-    };
-
-    pub fn new(max_requests: u64, window_secs: u64) -> Self {
-        Self {
-            limiters: std::sync::Mutex::new(lru::LruCache::new(Self::MAX_KEYS)),
-            max_requests,
-            window_secs,
-        }
-    }
-
-    /// Try to consume one request for the given user. Returns `true` if allowed.
-    pub fn check(&self, user_id: &str) -> bool {
-        let mut map = match self.limiters.lock() {
-            Ok(m) => m,
-            Err(e) => {
-                tracing::warn!("PerUserRateLimiter lock poisoned; recovering");
-                e.into_inner()
-            }
-        };
-        let limiter = map.get_or_insert_mut(user_id.to_string(), || {
-            RateLimiter::new(self.max_requests, self.window_secs)
-        });
-        limiter.check()
-    }
-}
-
-/// Per-user workspace pool: lazily creates and caches workspaces keyed by user_id.
-///
-/// In single-user mode, exactly one workspace is cached. In multi-user mode,
-/// each authenticated user gets their own workspace with appropriate scopes,
-/// search config, memory layers, and embedding cache settings.
-///
-/// Also implements [`WorkspaceResolver`] so it can be shared with memory tools,
-/// avoiding a separate `PerUserWorkspaceResolver` with duplicated logic.
-pub struct WorkspacePool {
-    db: Arc<dyn Database>,
-    embeddings: Option<Arc<dyn crate::workspace::EmbeddingProvider>>,
-    embedding_cache_config: crate::workspace::EmbeddingCacheConfig,
-    search_config: crate::config::WorkspaceSearchConfig,
-    workspace_config: crate::config::WorkspaceConfig,
-    cache: tokio::sync::RwLock<std::collections::HashMap<String, Arc<Workspace>>>,
-    /// Cached admin system prompt content. `None` = not yet loaded;
-    /// `Some("")` = loaded but empty/not set.
-    admin_prompt_cache: Arc<tokio::sync::RwLock<Option<String>>>,
-}
-
-impl WorkspacePool {
-    pub fn new(
-        db: Arc<dyn Database>,
-        embeddings: Option<Arc<dyn crate::workspace::EmbeddingProvider>>,
-        embedding_cache_config: crate::workspace::EmbeddingCacheConfig,
-        search_config: crate::config::WorkspaceSearchConfig,
-        workspace_config: crate::config::WorkspaceConfig,
-    ) -> Self {
-        Self {
-            db,
-            embeddings,
-            embedding_cache_config,
-            search_config,
-            workspace_config,
-            cache: tokio::sync::RwLock::new(std::collections::HashMap::new()),
-            admin_prompt_cache: Arc::new(tokio::sync::RwLock::new(None)),
-        }
-    }
-
-    /// Clear the admin prompt cache. Called after the PUT handler updates
-    /// the prompt so all workspaces see the new content on the next turn.
-    pub async fn invalidate_admin_prompt(&self) {
-        let mut guard = self.admin_prompt_cache.write().await;
-        *guard = None;
-    }
-
-    /// Build a workspace for a user, applying search config, embeddings,
-    /// global read scopes, memory layers, and admin prompt.
-    fn build_workspace(&self, user_id: &str) -> Workspace {
-        let mut ws = Workspace::new_with_db(user_id, Arc::clone(&self.db))
-            .with_search_config(&self.search_config)
-            .with_admin_prompt()
-            .with_admin_prompt_cache(Arc::clone(&self.admin_prompt_cache));
-
-        if let Some(ref emb) = self.embeddings {
-            ws = ws.with_embeddings_cached(Arc::clone(emb), self.embedding_cache_config.clone());
-        }
-
-        if !self.workspace_config.read_scopes.is_empty() {
-            ws = ws.with_additional_read_scopes(self.workspace_config.read_scopes.clone());
-        }
-
-        ws = ws.with_memory_layers(self.workspace_config.memory_layers.clone());
-        ws
-    }
-
-    /// Get or create a workspace for the given user identity.
-    ///
-    /// Applies search config, memory layers, embedding cache, and read scopes
-    /// (both from global config and from the token's `workspace_read_scopes`).
-    pub async fn get_or_create(&self, identity: &UserIdentity) -> Arc<Workspace> {
-        // Fast path: check read lock
-        {
-            let cache = self.cache.read().await;
-            if let Some(ws) = cache.get(&identity.user_id) {
-                return Arc::clone(ws);
-            }
-        }
-
-        // Slow path: create workspace under write lock
-        let mut cache = self.cache.write().await;
-        // Double-check after acquiring write lock
-        if let Some(ws) = cache.get(&identity.user_id) {
-            return Arc::clone(ws);
-        }
-
-        let mut ws = self.build_workspace(&identity.user_id);
-
-        // Apply per-token read scopes from identity.
-        if !identity.workspace_read_scopes.is_empty() {
-            ws = ws.with_additional_read_scopes(identity.workspace_read_scopes.clone());
-        }
-
-        let ws = Arc::new(ws);
-
-        cache.insert(identity.user_id.clone(), Arc::clone(&ws));
-
-        // Seed identity files after inserting into cache (so the lock can be
-        // dropped) but before returning, so callers see a seeded workspace.
-        // Drop the write lock explicitly before the async seed to avoid
-        // blocking other workspace lookups.
-        drop(cache);
-        if let Err(e) = ws.seed_if_empty().await {
-            tracing::warn!(
-                user_id = identity.user_id,
-                "Failed to seed workspace: {}",
-                e
-            );
-        }
-
-        ws
-    }
-}
-
-#[async_trait::async_trait]
-impl crate::tools::builtin::memory::WorkspaceResolver for WorkspacePool {
-    async fn resolve(&self, user_id: &str) -> Arc<Workspace> {
-        // Fast path: check read lock
-        {
-            let cache = self.cache.read().await;
-            if let Some(ws) = cache.get(user_id) {
-                return Arc::clone(ws);
-            }
-        }
-
-        // Slow path: create workspace under write lock
-        let mut cache = self.cache.write().await;
-        if let Some(ws) = cache.get(user_id) {
-            return Arc::clone(ws);
-        }
-
-        let ws = Arc::new(self.build_workspace(user_id));
-        cache.insert(user_id.to_string(), Arc::clone(&ws));
-        drop(cache);
-
-        // Match the seeded workspace behavior used by the prompt-side lookup so
-        // v1 memory tools and the v1 system prompt see the same per-user scope.
-        if let Err(e) = ws.seed_if_empty().await {
-            tracing::warn!(user_id = user_id, "Failed to seed workspace: {}", e);
-        }
-
-        tracing::debug!(user_id = user_id, "Created per-user workspace");
-        ws
-    }
-}
-
-/// Shared state for all gateway handlers.
-pub struct GatewayState {
-    /// Channel to send messages to the agent loop.
-    pub msg_tx: tokio::sync::RwLock<Option<mpsc::Sender<IncomingMessage>>>,
-    /// SSE broadcast manager (Arc-wrapped so extension manager can hold a reference).
-    pub sse: Arc<SseManager>,
-    /// Workspace for memory API (single-user fallback).
-    pub workspace: Option<Arc<Workspace>>,
-    /// Per-user workspace pool for multi-user mode.
-    pub workspace_pool: Option<Arc<WorkspacePool>>,
-    /// Session manager for thread info.
-    pub session_manager: Option<Arc<SessionManager>>,
-    /// Log broadcaster for the logs SSE endpoint.
-    pub log_broadcaster: Option<Arc<LogBroadcaster>>,
-    /// Handle for changing the tracing log level at runtime.
-    pub log_level_handle: Option<Arc<crate::channels::web::log_layer::LogLevelHandle>>,
-    /// Extension manager for extension management API.
-    pub extension_manager: Option<Arc<ExtensionManager>>,
-    /// Tool registry for listing registered tools.
-    pub tool_registry: Option<Arc<ToolRegistry>>,
-    /// Database store for sandbox job persistence.
-    pub store: Option<Arc<dyn Database>>,
-    /// Cached settings store. When present, settings reads/writes go through
-    /// the cache layer for consistency with the agent loop. Concrete type so
-    /// handlers can also call `invalidate_user()` / `flush()`.
-    pub settings_cache: Option<Arc<crate::db::cached_settings::CachedSettingsStore>>,
-    /// Container job manager for sandbox operations.
-    pub job_manager: Option<Arc<ContainerJobManager>>,
-    /// Prompt queue for Claude Code follow-up prompts.
-    pub prompt_queue: Option<PromptQueue>,
-    /// Durable owner scope for persistence and unauthenticated callback flows.
-    pub owner_id: String,
-    /// Shutdown signal sender.
-    pub shutdown_tx: tokio::sync::RwLock<Option<oneshot::Sender<()>>>,
-    /// WebSocket connection tracker.
-    pub ws_tracker: Option<Arc<crate::channels::web::ws::WsConnectionTracker>>,
-    /// LLM provider for OpenAI-compatible API proxy.
-    pub llm_provider: Option<Arc<dyn crate::llm::LlmProvider>>,
-    /// Skill registry for skill management API.
-    pub skill_registry: Option<Arc<std::sync::RwLock<ironclaw_skills::SkillRegistry>>>,
-    /// Skill catalog for searching the ClawHub registry.
-    pub skill_catalog: Option<Arc<ironclaw_skills::catalog::SkillCatalog>>,
-    /// Shared auth manager for gateway auth submission and readiness checks.
-    pub auth_manager: Option<Arc<crate::bridge::auth_manager::AuthManager>>,
-    /// Scheduler for sending follow-up messages to running agent jobs.
-    pub scheduler: Option<crate::tools::builtin::SchedulerSlot>,
-    /// Per-user rate limiter for chat endpoints (30 messages per 60 seconds per user).
-    pub chat_rate_limiter: PerUserRateLimiter,
-    /// Per-IP rate limiter for OAuth/auth endpoints (20 requests per 60 seconds per IP).
-    pub oauth_rate_limiter: PerUserRateLimiter,
-    /// Rate limiter for webhook trigger endpoints (10 requests per 60 seconds).
-    pub webhook_rate_limiter: RateLimiter,
-    /// Registry catalog entries for the available extensions API.
-    /// Populated at startup from `registry/` manifests, independent of extension manager.
-    pub registry_entries: Vec<crate::extensions::RegistryEntry>,
-    /// Cost guard for token/cost tracking.
-    pub cost_guard: Option<Arc<crate::agent::cost_guard::CostGuard>>,
-    /// Routine engine slot for manual routine triggering (filled at runtime).
-    pub routine_engine: RoutineEngineSlot,
-    /// Server startup time for uptime calculation.
-    pub startup_time: std::time::Instant,
-    /// Snapshot of active (resolved) configuration for the frontend.
-    pub active_config: ActiveConfigSnapshot,
-    /// Secrets store for admin secret provisioning.
-    pub secrets_store: Option<Arc<dyn crate::secrets::SecretsStore + Send + Sync>>,
-    /// DB auth cache for invalidation on security-critical actions.
-    pub db_auth: Option<Arc<crate::channels::web::auth::DbAuthenticator>>,
-    /// Shared pairing store (one instance per server, not per request).
-    pub pairing_store: Option<Arc<crate::pairing::PairingStore>>,
-    /// OAuth providers for social login (None when OAuth is disabled).
-    pub oauth_providers: Option<
-        Arc<
-            std::collections::HashMap<
-                String,
-                Arc<dyn crate::channels::web::oauth::providers::OAuthProvider>,
-            >,
-        >,
-    >,
-    /// In-memory store for pending OAuth flows (CSRF + PKCE state).
-    pub oauth_state_store: Option<Arc<crate::channels::web::oauth::state_store::OAuthStateStore>>,
-    /// Base URL for constructing OAuth callback URLs.
-    pub oauth_base_url: Option<String>,
-    /// Email domains allowed for OAuth/OIDC login. Empty means allow all.
-    pub oauth_allowed_domains: Vec<String>,
-    /// NEAR wallet auth nonce store (None when NEAR auth is disabled).
-    pub near_nonce_store: Option<Arc<crate::channels::web::oauth::near::NearNonceStore>>,
-    /// NEAR RPC endpoint URL for access key verification.
-    pub near_rpc_url: Option<String>,
-    /// NEAR network name (mainnet/testnet) for the frontend wallet connector.
-    pub near_network: Option<String>,
-    /// Shutdown signal for OAuth/NEAR sweep background tasks.
-    /// When this sender is dropped, the sweep loops exit gracefully.
-    #[allow(dead_code)]
-    pub oauth_sweep_shutdown: Option<tokio::sync::watch::Sender<()>>,
-    /// Cache for the assembled frontend HTML served from `/`.
-    ///
-    /// The cache key is derived from the `updated_at` of
-    /// `.system/gateway/layout.json` and the `.system/gateway/widgets/`
-    /// directory — both returned by a single cheap `list(".system/gateway/")`
-    /// call. A hit skips reading the layout, every widget manifest, every
-    /// widget JS file, and every widget CSS file. A miss (or absent cache)
-    /// falls through to the full `build_frontend_html()` path.
-    pub frontend_html_cache: Arc<tokio::sync::RwLock<Option<FrontendHtmlCache>>>,
-    /// Channel-agnostic tool dispatcher for routing handler operations through
-    /// the tool pipeline with audit trail.
-    pub tool_dispatcher: Option<Arc<crate::tools::dispatch::ToolDispatcher>>,
-}
-
-/// Cached result of `build_frontend_html()`, keyed by a cheap workspace
-/// signature so the fast path only needs one `list()` call per request.
-#[derive(Debug, Clone)]
-pub struct FrontendHtmlCache {
-    /// Signature the cache is valid for. The cache is bypassed when the
-    /// current workspace signature differs from this one.
-    pub key: FrontendCacheKey,
-    /// The assembled HTML, or `None` if the layout had no customizations
-    /// and the caller should serve the embedded default unchanged.
-    pub html: Option<String>,
-}
-
-/// Cheap workspace fingerprint covering the inputs of `build_frontend_html`.
-///
-/// Uses the per-entry `updated_at` timestamps returned by `Workspace::list`
-/// (the directory entry's `updated_at` is "latest among children", so widget
-/// file edits bubble up automatically). Timestamps are stored as
-/// `(seconds, nanoseconds)` pairs to avoid depending on `chrono` types here.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct FrontendCacheKey {
-    /// Signature for `.system/gateway/layout.json`, or `None` if absent.
-    pub layout: Option<(i64, u32)>,
-    /// Signature for `.system/gateway/widgets/` (max child mtime), or `None`
-    /// if the directory is empty or absent.
-    pub widgets: Option<(i64, u32)>,
 }
 
 /// Start the gateway HTTP server.
@@ -977,606 +556,6 @@ pub async fn start_server(
     });
 
     Ok(bound_addr)
-}
-
-// --- Content Security Policy ---
-//
-// A single source of truth for the gateway's CSP. The static value below is
-// used by the global response-header layer for every endpoint. The
-// ---- Content-Security-Policy construction ----------------------------
-//
-// The gateway serves two flavors of CSP on the same set of directives:
-//
-// * The static header applied by `SetResponseHeaderLayer` to *every*
-//   response (see [`BASE_CSP_HEADER`]). No inline scripts are authorized.
-// * A per-response variant produced by [`build_csp`] with a `'nonce-…'`
-//   source added to `script-src`, used only by `index_handler` when it
-//   serves customized HTML containing inline `<script>` blocks.
-//
-// Both variants MUST carry the same directive set except for `script-src`
-// — if one grows a new `connect-src` origin, the other silently stays on
-// the old policy, and customized pages end up under a stricter CSP than
-// plain pages (or vice versa). Previous versions of this file duplicated
-// the full directive string in two places, so adding a CDN to one was a
-// latent regression waiting to happen. Keep every directive as a named
-// constant and assemble both flavors via [`build_csp`] so there is a
-// single source of truth.
-
-/// `script-src` sources other than `'self'` and the per-response nonce.
-const SCRIPT_SRC_EXTRAS: &str =
-    "https://cdn.jsdelivr.net https://cdnjs.cloudflare.com https://esm.sh";
-const STYLE_SRC: &str = "'self' 'unsafe-inline' https://fonts.googleapis.com";
-const FONT_SRC: &str = "https://fonts.gstatic.com data:";
-const CONNECT_SRC: &str =
-    "'self' https://esm.sh https://rpc.mainnet.near.org https://rpc.testnet.near.org";
-const IMG_SRC: &str =
-    "'self' data: blob: https://*.googleusercontent.com https://avatars.githubusercontent.com";
-const FRAME_SRC: &str = "https://accounts.google.com https://appleid.apple.com";
-const FORM_ACTION: &str =
-    "'self' https://accounts.google.com https://github.com https://appleid.apple.com";
-
-/// Build a CSP string. When `nonce` is `Some`, the resulting policy adds
-/// `'nonce-{nonce}'` to `script-src` so a single inline `<script
-/// nonce="{nonce}">` block on the same response is authorized. When
-/// `nonce` is `None`, the policy matches the static header emitted by
-/// [`BASE_CSP_HEADER`]. This is the single source of truth for the
-/// gateway CSP — edit per-directive constants above, not the format
-/// string here.
-fn build_csp(nonce: Option<&str>) -> String {
-    let script_nonce = match nonce {
-        Some(n) => format!(" 'nonce-{n}'"),
-        None => String::new(),
-    };
-    format!(
-        "default-src 'self'; \
-         script-src 'self'{script_nonce} {SCRIPT_SRC_EXTRAS}; \
-         style-src {STYLE_SRC}; \
-         font-src {FONT_SRC}; \
-         connect-src {CONNECT_SRC}; \
-         img-src {IMG_SRC}; \
-         frame-src {FRAME_SRC}; \
-         object-src 'none'; \
-         frame-ancestors 'none'; \
-         base-uri 'self'; \
-         form-action {FORM_ACTION}"
-    )
-}
-
-/// Static CSP header applied to every gateway response by the
-/// response-header layer. Assembled at first use via [`build_csp`] with no
-/// nonce. Falls back to a minimally-permissive `default-src 'self'` if the
-/// assembled value somehow fails to parse as a `HeaderValue` — in practice
-/// the assembled string is pure ASCII and this branch is unreachable, but
-/// production code in this repo doesn't use `.expect()` on request-path
-/// values.
-static BASE_CSP_HEADER: std::sync::LazyLock<header::HeaderValue> = std::sync::LazyLock::new(|| {
-    header::HeaderValue::from_str(&build_csp(None))
-        .unwrap_or_else(|_| header::HeaderValue::from_static("default-src 'self'"))
-});
-
-/// Build a CSP equivalent to the static header but with `'nonce-{nonce}'`
-/// added to the `script-src` directive. Thin wrapper kept for call-site
-/// readability (the name is the contract the nonce handler wants).
-fn build_csp_with_nonce(nonce: &str) -> String {
-    build_csp(Some(nonce))
-}
-
-/// Generate a fresh per-response CSP nonce. 16 random bytes hex-encoded
-/// (32 chars) — well above the 128-bit minimum recommended for nonces and
-/// matching the `OsRng + hex` pattern used elsewhere in this module
-/// (see `tokens_create_handler`).
-fn generate_csp_nonce() -> String {
-    use rand::RngCore;
-    use rand::rngs::OsRng;
-    let mut bytes = [0u8; 16];
-    OsRng.fill_bytes(&mut bytes);
-    hex::encode(bytes)
-}
-
-// --- Frontend bundle assembly ---
-
-use ironclaw_gateway::assets;
-use ironclaw_gateway::{FrontendBundle, LayoutConfig, NONCE_PLACEHOLDER};
-
-use crate::channels::web::handlers::frontend::{load_resolved_widgets, read_layout_config};
-
-/// Compute a cheap cache key for `build_frontend_html` — one `list` call
-/// against `.system/gateway/`. The directory entry for `widgets/` carries the
-/// max `updated_at` of its children, so any widget file edit naturally bubbles
-/// into the key without needing to read individual manifests.
-async fn compute_frontend_cache_key(workspace: &crate::workspace::Workspace) -> FrontendCacheKey {
-    let Ok(entries) = workspace.list(".system/gateway/").await else {
-        return FrontendCacheKey {
-            layout: None,
-            widgets: None,
-        };
-    };
-    let mut key = FrontendCacheKey {
-        layout: None,
-        widgets: None,
-    };
-    for entry in entries {
-        let ts = entry
-            .updated_at
-            .map(|t| (t.timestamp(), t.timestamp_subsec_nanos()));
-        match entry.name() {
-            "layout.json" if !entry.is_directory => key.layout = ts,
-            "widgets" if entry.is_directory => key.widgets = ts,
-            _ => {}
-        }
-    }
-    key
-}
-
-/// Build customized HTML from the workspace gateway config.
-///
-/// Returns `None` if the workspace is unavailable or the loaded layout has no
-/// customizations and no widgets — in that case the caller serves the embedded
-/// default HTML unchanged. Custom CSS is deliberately **not** included in the
-/// returned bundle: `css_handler` appends `.system/gateway/custom.css` onto
-/// `/style.css` so the stylesheet is the single source of truth for CSS
-/// overrides.
-///
-/// The assembled HTML is cached in `GatewayState::frontend_html_cache` behind
-/// a fingerprint of `.system/gateway/layout.json` and `.system/gateway/widgets/`
-/// mtimes (computed with a single `list()` call). A cache hit skips reading
-/// every widget manifest / JS / CSS file, which would otherwise fire on every
-/// page load.
-///
-/// **Multi-tenant safety.** In multi-user mode (`workspace_pool` set) this
-/// function ALWAYS returns `None`, regardless of whether `state.workspace` is
-/// also populated. The customization assembly path is fundamentally
-/// single-tenant: `index_handler` (`GET /`) is the unauthenticated bootstrap
-/// route — no user identity is available at request time, so there is no way
-/// to resolve the *correct* per-user workspace inside this function. Reading
-/// `state.workspace` instead would expose one global workspace's
-/// customizations to every user, and the process-wide
-/// `frontend_html_cache` would pin the leak across requests. We refuse the
-/// path entirely and serve the embedded default to all users; per-user
-/// customization can ride a future JS-side fetch against
-/// `/api/frontend/layout`, which is authenticated and routes through
-/// `resolve_workspace(&state, &user)` so it returns the right workspace.
-/// See `crates/ironclaw_gateway/static/app.js` — the layout-config IIFE
-/// already reads `window.__IRONCLAW_LAYOUT__`, which a future change can
-/// populate from a `fetch('/api/frontend/layout')` after auth.
-///
-/// **Cache key TOCTOU window (known and accepted).** The fast-path cache
-/// key is computed by [`compute_frontend_cache_key`] in a single
-/// `Workspace::list` call, but the slow-path data read
-/// (`read_layout_config` + `load_resolved_widgets`) happens *after* that
-/// key is observed, in separate workspace operations. A workspace write
-/// landing between the two — operator edits `layout.json` while a
-/// request is mid-rebuild — can therefore produce a cache entry whose
-/// HTML was assembled from a layout *newer* than the key it's stored
-/// under. The next request after the writes settle will recompute the
-/// key, see a different fingerprint, and replace the cache entry, so
-/// the staleness window is always self-correcting and bounded by one
-/// rebuild round-trip.
-///
-/// This is intentional. Making the read+key+store sequence atomic would
-/// require a workspace-level read lock that the rest of the gateway
-/// doesn't take, and would punish the (much hotter) cache hit path with
-/// extra coordination. The acceptability rests on three observations:
-/// (a) the staleness window is bounded by a single `list()` call's
-/// worth of wall time, (b) the cache is per-process so the staleness
-/// can never outlive `Drop` of `GatewayState`, and (c) layout writes
-/// are rare and operator-initiated — there is no realistic workload
-/// that fires a write at the cadence required to keep the entry
-/// permanently stale. If a future workload changes that calculus, the
-/// right fix is a workspace version generation counter, not a lock
-/// around this function.
-async fn build_frontend_html(state: &GatewayState) -> Option<String> {
-    if state.workspace_pool.is_some() {
-        // Multi-tenant: refuse the assembly path entirely. See the function
-        // doc comment above for the full rationale. The cache write below
-        // is unreachable on this branch, so the cache stays empty and
-        // cannot leak one user's customizations to another.
-        return None;
-    }
-
-    let ws = state.workspace.as_ref()?;
-
-    // Fast path — cache hit. One workspace `list()` call, no file reads.
-    let cache_key = compute_frontend_cache_key(ws).await;
-    {
-        let cache = state.frontend_html_cache.read().await;
-        if let Some(ref cached) = *cache
-            && cached.key == cache_key
-        {
-            return cached.html.clone();
-        }
-    }
-
-    // Slow path — rebuild.
-    let layout = read_layout_config(ws).await;
-    let widgets = load_resolved_widgets(ws, &layout).await;
-
-    // Skip assembly when nothing is customized. `layout_has_customizations`
-    // is the single source of truth so adding a new field to `LayoutConfig`
-    // forces an update in one place instead of a big boolean expression here.
-    let html = if widgets.is_empty() && !layout_has_customizations(&layout) {
-        None
-    } else {
-        let bundle = FrontendBundle {
-            layout,
-            widgets,
-            // Custom CSS is served via /style.css (css_handler) to avoid
-            // double-application — see the doc comment on this function.
-            custom_css: None,
-        };
-        Some(ironclaw_gateway::assemble_index(
-            assets::INDEX_HTML,
-            &bundle,
-        ))
-    };
-
-    // Store in cache. If another request raced us here, either writer wins —
-    // both produced the same HTML for the same key, so the cache ends up
-    // consistent either way.
-    *state.frontend_html_cache.write().await = Some(FrontendHtmlCache {
-        key: cache_key,
-        html: html.clone(),
-    });
-
-    html
-}
-
-/// Returns `true` if the layout config has any field that would affect the
-/// rendered HTML. When this returns `false` and there are no widgets, the
-/// gateway serves the embedded default unchanged.
-fn layout_has_customizations(layout: &LayoutConfig) -> bool {
-    let b = &layout.branding;
-    let t = &layout.tabs;
-    let c = &layout.chat;
-    // `branding.colors` is opaque to this function — `BrandingColors` may
-    // exist as `Some({})` (both fields `None`) or with values that the
-    // `is_safe_css_color` validator strips at injection time. Treating
-    // bare `colors.is_some()` as a customization forces the customized
-    // HTML path (and the per-response nonce CSP that comes with it) for
-    // layouts that produce zero effective branding output. Require at
-    // least one trimmed-non-empty color field, mirroring what
-    // `to_css_vars` actually emits.
-    let has_branding_colors = b.colors.as_ref().is_some_and(|colors| {
-        let nonempty = |v: &Option<String>| v.as_deref().is_some_and(|s| !s.trim().is_empty());
-        nonempty(&colors.primary) || nonempty(&colors.accent)
-    });
-    // Same precedent for URL fields: route through the `safe_logo_url`
-    // / `safe_favicon_url` getters that apply `is_safe_url`. A
-    // `layout.json` with `logo_url: "javascript:alert(1)"` would
-    // otherwise force the customized HTML path even though the value
-    // gets dropped at consumer time. Symmetric with how branding colors
-    // are gated above.
-    b.title.is_some()
-        || b.subtitle.is_some()
-        || b.safe_logo_url().is_some()
-        || b.safe_favicon_url().is_some()
-        || has_branding_colors
-        || t.order.is_some()
-        || t.hidden.is_some()
-        || t.default_tab.is_some()
-        || c.suggestions.is_some()
-        || c.image_upload.is_some()
-        || c.upgrade_inline_json.is_some()
-        || !layout.widgets.is_empty()
-}
-
-// --- Static file handlers ---
-//
-// All frontend assets are embedded in the `ironclaw_gateway` crate.
-// These handlers serve them with appropriate MIME types and cache headers.
-
-/// Substitute [`NONCE_PLACEHOLDER`] sentinels in the assembled HTML with a
-/// fresh per-response CSP nonce.
-///
-/// **Why an attribute-targeted replace, not a bare string replace.** The
-/// assembled HTML embeds widget JavaScript inline (so a CSP-protected
-/// `<script src>` doesn't need to authenticate against `/api/frontend/widget/...`).
-/// A widget author has every right to write the literal string
-/// `__IRONCLAW_CSP_NONCE__` inside their own source — in a comment, a log
-/// line, a test fixture, or just as a constant they happen to define. A
-/// naive `html.replace(NONCE_PLACEHOLDER, nonce)` would silently rewrite
-/// every such occurrence into a per-request nonce, mutating widget code
-/// in a way the author didn't ask for.
-///
-/// The substitution here targets the full attribute form
-/// `nonce="__IRONCLAW_CSP_NONCE__"`, which is the exact shape
-/// `assemble_index` emits when stamping nonces onto `<script>` tags. The
-/// double-quoted sentinel is unambiguous in HTML context — it can never
-/// accidentally match free text in a JS module body, a comment, or a
-/// JSON payload. Inline `<style>` blocks deliberately get no nonce
-/// (style-src allows `'unsafe-inline'`) so they're untouched either way.
-fn stamp_nonce_into_html(html_with_placeholder: &str, nonce: &str) -> String {
-    let placeholder_attr = format!("nonce=\"{NONCE_PLACEHOLDER}\"");
-    let nonce_attr = format!("nonce=\"{nonce}\"");
-    html_with_placeholder.replace(&placeholder_attr, &nonce_attr)
-}
-
-async fn index_handler(State(state): State<Arc<GatewayState>>) -> Response {
-    // Try to assemble customized HTML from workspace frontend config.
-    // Falls back to embedded HTML if workspace is unavailable or has no
-    // customizations — in that case there are no inline scripts and the
-    // global CSP layer applies unchanged.
-    let assembled = build_frontend_html(&state).await;
-
-    let Some(html_with_placeholder) = assembled else {
-        return (
-            [
-                (header::CONTENT_TYPE, "text/html; charset=utf-8"),
-                (header::CACHE_CONTROL, "no-cache"),
-            ],
-            assets::INDEX_HTML,
-        )
-            .into_response();
-    };
-
-    // Customized path: the assembled HTML contains inline `<script>` blocks
-    // (layout config + widget modules) carrying [`NONCE_PLACEHOLDER`] in
-    // their `nonce` attribute. Stamp a fresh per-response nonce in both
-    // the HTML and the response's Content-Security-Policy header so the
-    // browser actually executes the scripts.
-    //
-    // Setting `Content-Security-Policy` here suppresses the global
-    // `SetResponseHeaderLayer::if_not_present` value for this response only.
-    let nonce = generate_csp_nonce();
-    let html = stamp_nonce_into_html(&html_with_placeholder, &nonce);
-    let csp = build_csp_with_nonce(&nonce);
-
-    (
-        [
-            (header::CONTENT_TYPE, "text/html; charset=utf-8".to_string()),
-            (header::CACHE_CONTROL, "no-cache".to_string()),
-            (
-                header::HeaderName::from_static("content-security-policy"),
-                csp,
-            ),
-        ],
-        html,
-    )
-        .into_response()
-}
-
-/// Compute the strong ETag value for a CSS body.
-///
-/// Strong validators are quoted, sha-prefixed, and truncated to 16 hex chars
-/// (64 bits) — collisions are statistically irrelevant for cache validation
-/// and the short form keeps headers compact. The same scheme is used for
-/// both the embedded base stylesheet and the workspace-customized variant
-/// so a flip between the two flavors naturally invalidates the client's
-/// cached copy.
-fn css_etag(body: &str) -> String {
-    let digest = Sha256::digest(body.as_bytes());
-    let hex = hex::encode(digest);
-    // 16 hex chars = 64 bits, plenty for content addressing.
-    format!("\"sha256-{}\"", &hex[..16])
-}
-
-async fn css_handler(State(state): State<Arc<GatewayState>>, headers: HeaderMap) -> Response {
-    // Append custom CSS from `.system/gateway/custom.css` if it exists.
-    //
-    // The hot path (no workspace overlay) borrows `assets::STYLE_CSS` directly
-    // via `Cow::Borrowed` so we don't allocate / copy the entire embedded
-    // stylesheet on every request. We only fall through to an owned
-    // `format!` when there's actually content to append.
-    //
-    // **Multi-tenant safety.** This must mirror the same guard
-    // `build_frontend_html` already enforces (see its doc comment): in
-    // multi-user mode (`workspace_pool.is_some()`) we cannot resolve a
-    // per-user workspace because `/style.css` is the unauthenticated
-    // bootstrap stylesheet — there is no user identity at request time.
-    // Reading from `state.workspace` here would expose one global
-    // workspace's `custom.css` to every user, defeating the
-    // `index_handler` guard at the sibling endpoint. Refuse the overlay
-    // path entirely in multi-tenant mode and serve the embedded base
-    // stylesheet to all users; per-user CSS overrides can ride a future
-    // authenticated `/api/frontend/custom-css` endpoint.
-    let css: std::borrow::Cow<'static, str> = if state.workspace_pool.is_some() {
-        std::borrow::Cow::Borrowed(assets::STYLE_CSS)
-    } else {
-        match &state.workspace {
-            Some(ws) => match ws.read(".system/gateway/custom.css").await {
-                Ok(doc) if !doc.content.trim().is_empty() => std::borrow::Cow::Owned(format!(
-                    "{}\n/* --- custom overrides --- */\n{}",
-                    assets::STYLE_CSS,
-                    doc.content
-                )),
-                _ => std::borrow::Cow::Borrowed(assets::STYLE_CSS),
-            },
-            None => std::borrow::Cow::Borrowed(assets::STYLE_CSS),
-        }
-    };
-
-    // Strong validator over the assembled body. The cache key naturally
-    // tracks both base stylesheet edits (compile-time) and `custom.css`
-    // edits (workspace mutation) — operators no longer need to ask users
-    // to hard-refresh after tweaking branding.
-    let etag = css_etag(&css);
-
-    // Conditional GET: if the client already holds this exact body, send a
-    // 304 with no body and let the browser reuse its cached copy. RFC 9110
-    // §13.1.2 — `If-None-Match` is a list of validators; we accept either
-    // an exact match or the literal `*`. Anything else falls through to a
-    // full 200 response.
-    if let Some(value) = headers.get(header::IF_NONE_MATCH)
-        && let Ok(s) = value.to_str()
-        && s.split(',').any(|v| {
-            let v = v.trim();
-            v == "*" || v == etag
-        })
-    {
-        return (
-            StatusCode::NOT_MODIFIED,
-            [
-                (header::ETAG, etag.as_str()),
-                (header::CACHE_CONTROL, "no-cache"),
-            ],
-        )
-            .into_response();
-    }
-
-    (
-        [
-            (header::CONTENT_TYPE, "text/css".to_string()),
-            // Keep `no-cache` so the browser always revalidates — combined
-            // with the ETag this gives us "fast 304" semantics rather than
-            // a stale `max-age` window where operator edits don't show up.
-            (header::CACHE_CONTROL, "no-cache".to_string()),
-            (header::ETAG, etag),
-        ],
-        css,
-    )
-        .into_response()
-}
-
-async fn theme_css_handler() -> impl IntoResponse {
-    (
-        [
-            (header::CONTENT_TYPE, "text/css"),
-            (header::CACHE_CONTROL, "no-cache"),
-        ],
-        assets::THEME_CSS,
-    )
-}
-
-async fn js_handler() -> impl IntoResponse {
-    (
-        [
-            (header::CONTENT_TYPE, "application/javascript"),
-            (header::CACHE_CONTROL, "no-cache"),
-        ],
-        assets::APP_JS,
-    )
-}
-
-async fn theme_init_handler() -> impl IntoResponse {
-    (
-        [
-            (header::CONTENT_TYPE, "application/javascript"),
-            (header::CACHE_CONTROL, "no-cache"),
-        ],
-        assets::THEME_INIT_JS,
-    )
-}
-
-async fn favicon_handler() -> impl IntoResponse {
-    (
-        [
-            (header::CONTENT_TYPE, "image/x-icon"),
-            (header::CACHE_CONTROL, "public, max-age=86400"),
-        ],
-        assets::FAVICON_ICO,
-    )
-}
-
-async fn i18n_index_handler() -> impl IntoResponse {
-    (
-        [
-            (header::CONTENT_TYPE, "application/javascript"),
-            (header::CACHE_CONTROL, "no-cache"),
-        ],
-        assets::I18N_INDEX_JS,
-    )
-}
-
-async fn i18n_en_handler() -> impl IntoResponse {
-    (
-        [
-            (header::CONTENT_TYPE, "application/javascript"),
-            (header::CACHE_CONTROL, "no-cache"),
-        ],
-        assets::I18N_EN_JS,
-    )
-}
-
-async fn i18n_zh_handler() -> impl IntoResponse {
-    (
-        [
-            (header::CONTENT_TYPE, "application/javascript"),
-            (header::CACHE_CONTROL, "no-cache"),
-        ],
-        assets::I18N_ZH_CN_JS,
-    )
-}
-
-async fn i18n_ko_handler() -> impl IntoResponse {
-    (
-        [
-            (header::CONTENT_TYPE, "application/javascript"),
-            (header::CACHE_CONTROL, "no-cache"),
-        ],
-        assets::I18N_KO_JS,
-    )
-}
-
-async fn i18n_app_handler() -> impl IntoResponse {
-    (
-        [
-            (header::CONTENT_TYPE, "application/javascript"),
-            (header::CACHE_CONTROL, "no-cache"),
-        ],
-        assets::I18N_APP_JS,
-    )
-}
-
-// --- Admin panel static handlers ---
-
-async fn admin_html_handler() -> impl IntoResponse {
-    // Admin panel CSP — fully same-origin, no CDN allowances.
-    // Delivered as an HTTP header (not a <meta> tag) so the browser enforces
-    // it before any markup is parsed.
-    const ADMIN_CSP: &str = "default-src 'self'; \
-        script-src 'self'; \
-        style-src 'self' 'unsafe-inline'; \
-        font-src 'self'; \
-        connect-src 'self'; \
-        img-src 'self' data:; \
-        object-src 'none'; \
-        frame-ancestors 'none'; \
-        base-uri 'self'; \
-        form-action 'self'";
-
-    let mut headers = HeaderMap::new();
-    headers.insert(
-        header::CONTENT_TYPE,
-        header::HeaderValue::from_static("text/html; charset=utf-8"),
-    );
-    headers.insert(
-        header::CACHE_CONTROL,
-        header::HeaderValue::from_static("no-cache"),
-    );
-    headers.insert(
-        header::HeaderName::from_static("content-security-policy"),
-        header::HeaderValue::from_static(ADMIN_CSP),
-    );
-    (headers, assets::ADMIN_HTML)
-}
-
-async fn admin_css_handler() -> impl IntoResponse {
-    (
-        [
-            (header::CONTENT_TYPE, "text/css"),
-            (header::CACHE_CONTROL, "no-cache"),
-        ],
-        assets::ADMIN_CSS,
-    )
-}
-
-async fn admin_js_handler() -> impl IntoResponse {
-    (
-        [
-            (header::CONTENT_TYPE, "application/javascript"),
-            (header::CACHE_CONTROL, "no-cache"),
-        ],
-        assets::ADMIN_JS,
-    )
-}
-
-// --- Health ---
-
-async fn health_handler() -> Json<HealthResponse> {
-    Json(HealthResponse {
-        status: "healthy",
-        channel: "gateway",
-    })
 }
 
 /// Return an OAuth error landing page response.
@@ -3914,101 +2893,6 @@ async fn extensions_activate_handler(
     }
 }
 
-// --- Project file serving handlers ---
-
-/// Redirect `/projects/{id}` to `/projects/{id}/` so relative paths in
-/// the served HTML resolve within the project namespace.
-async fn project_redirect_handler(
-    State(state): State<Arc<GatewayState>>,
-    super::auth::AuthenticatedUser(user): super::auth::AuthenticatedUser,
-    Path(project_id): Path<String>,
-) -> impl IntoResponse {
-    if !verify_project_ownership(&state, &project_id, &user.user_id).await {
-        return (StatusCode::NOT_FOUND, "Not found").into_response();
-    }
-    axum::response::Redirect::permanent(&format!("/projects/{project_id}/")).into_response()
-}
-
-/// Serve `index.html` when hitting `/projects/{project_id}/`.
-async fn project_index_handler(
-    State(state): State<Arc<GatewayState>>,
-    super::auth::AuthenticatedUser(user): super::auth::AuthenticatedUser,
-    Path(project_id): Path<String>,
-) -> impl IntoResponse {
-    if !verify_project_ownership(&state, &project_id, &user.user_id).await {
-        return (StatusCode::NOT_FOUND, "Not found").into_response();
-    }
-    serve_project_file(&project_id, "index.html").await
-}
-
-/// Serve any file under `/projects/{project_id}/{path}`.
-async fn project_file_handler(
-    State(state): State<Arc<GatewayState>>,
-    super::auth::AuthenticatedUser(user): super::auth::AuthenticatedUser,
-    Path((project_id, path)): Path<(String, String)>,
-) -> impl IntoResponse {
-    if !verify_project_ownership(&state, &project_id, &user.user_id).await {
-        return (StatusCode::NOT_FOUND, "Not found").into_response();
-    }
-    serve_project_file(&project_id, &path).await
-}
-
-/// Check that a project directory belongs to a job owned by the given user.
-/// Returns false if the store is unavailable or the project is not found.
-async fn verify_project_ownership(state: &GatewayState, project_id: &str, user_id: &str) -> bool {
-    let Some(ref store) = state.store else {
-        return false;
-    };
-    // The project_id is a sandbox job UUID used as the directory name.
-    let Ok(job_id) = project_id.parse::<uuid::Uuid>() else {
-        return false;
-    };
-    match store.get_sandbox_job(job_id).await {
-        Ok(Some(job)) => job.user_id == user_id,
-        _ => false,
-    }
-}
-
-/// Shared logic: resolve the file inside `~/.ironclaw/projects/{project_id}/`,
-/// guard against path traversal, and stream the content with the right MIME type.
-async fn serve_project_file(project_id: &str, path: &str) -> axum::response::Response {
-    // Reject project_id values that could escape the projects directory.
-    if project_id.contains('/')
-        || project_id.contains('\\')
-        || project_id.contains("..")
-        || project_id.is_empty()
-    {
-        return (StatusCode::BAD_REQUEST, "Invalid project ID").into_response();
-    }
-
-    let base = ironclaw_base_dir().join("projects").join(project_id);
-
-    let file_path = base.join(path);
-
-    // Path traversal guard
-    let canonical = match file_path.canonicalize() {
-        Ok(p) => p,
-        Err(_) => return (StatusCode::NOT_FOUND, "Not found").into_response(),
-    };
-    let base_canonical = match base.canonicalize() {
-        Ok(p) => p,
-        Err(_) => return (StatusCode::NOT_FOUND, "Not found").into_response(),
-    };
-    if !canonical.starts_with(&base_canonical) {
-        return (StatusCode::FORBIDDEN, "Forbidden").into_response();
-    }
-
-    match tokio::fs::read(&canonical).await {
-        Ok(contents) => {
-            let mime = mime_guess::from_path(&canonical)
-                .first_or_octet_stream()
-                .to_string();
-            ([(header::CONTENT_TYPE, mime)], contents).into_response()
-        }
-        Err(_) => (StatusCode::NOT_FOUND, "Not found").into_response(),
-    }
-}
-
 async fn extensions_remove_handler(
     State(state): State<Arc<GatewayState>>,
     AuthenticatedUser(user): AuthenticatedUser,
@@ -4552,12 +3436,21 @@ mod tests {
     use super::*;
     use crate::agent::SessionManager;
     use crate::auth::oauth;
+    use crate::channels::web::auth::UserIdentity;
+    use crate::channels::web::platform::static_files::{
+        BASE_CSP_HEADER, build_csp, build_csp_with_nonce, build_frontend_html, css_etag,
+        generate_csp_nonce, stamp_nonce_into_html,
+    };
+    use crate::channels::web::sse::SseManager;
     use crate::channels::web::types::{
         ExtensionActivationStatus, classify_wasm_channel_activation,
     };
-    use crate::extensions::{ExtensionKind, InstalledExtension};
+    use crate::db::Database;
+    use crate::extensions::{ExtensionKind, ExtensionManager, InstalledExtension};
     use crate::testing::credentials::TEST_GATEWAY_CRYPTO_KEY;
-    use crate::tools::{Tool, ToolError, ToolOutput};
+    use crate::tools::{Tool, ToolError, ToolOutput, ToolRegistry};
+    use crate::workspace::Workspace;
+    use ironclaw_gateway::{NONCE_PLACEHOLDER, assets};
 
     #[test]
     fn test_build_turns_from_db_messages_complete() {
